Update babel dependencies and refactor tests to match Jest syntax.
