+++ .babelrc
@@ -1,1 +1,3 @@
-{ "presets": ["env"] }
+{
+  "presets": ["@babel/preset-env"]
+}
