+++ lib/__tests__/getLastWord.js
@@ -1,7 +1,7 @@
 import { Editor } from "slate";
 import Plain from "slate-plain-serializer";
 
-import { getLastWord } from "../lib/queries";
+import { getLastWord } from "../queries";
 
 describe("getLastWord()", () => {
   describe("when a single word is writen", () => {
@@ -9,10 +9,10 @@ describe("getLastWord()", () => {
       const editor = new Editor({ value: Plain.deserialize("hello") });
 
       // Anchor at the middle of the word
-      expect(getLastWord(editor.moveForward(2))).to.equal("he");
+      expect(getLastWord(editor.moveForward(2))).toBe("he");
 
       // Anchor at the end of the word
-      expect(getLastWord(editor.moveForward(3))).to.equal("hello");
+      expect(getLastWord(editor.moveForward(3))).toBe("hello");
     });
   });
 
@@ -21,13 +21,13 @@ describe("getLastWord()", () => {
       const editor = new Editor({ value: Plain.deserialize("hello world") });
 
       // Anchor at the end of the first word
-      expect(getLastWord(editor.moveForward(5))).to.equal("hello");
+      expect(getLastWord(editor.moveForward(5))).toBe("hello");
 
       // Anchor at the beginning of the second word
-      expect(getLastWord(editor.moveForward(1))).to.equal("");
+      expect(getLastWord(editor.moveForward(1))).toBe("");
 
       // Anchor at the end of the second word
-      expect(getLastWord(editor.moveForward(5))).to.equal("world");
+      expect(getLastWord(editor.moveForward(5))).toBe("world");
     });
   });
 });
