+++ lib/__tests__/index.js
@@ -1,7 +1,7 @@
 import { Editor } from "slate";
 import Plain from "slate-plain-serializer";
 
-import AutoReplacePlugin from "../lib";
+import AutoReplacePlugin from "../";
 
 const KeyDownInsertTextPlugin = textToInsert => ({
   onKeyDown(event, editor, next) {
@@ -54,7 +54,7 @@ describe("AutoReplacePlugin()", () => {
       simulator.run("onKeyDown", keyEvent("h"));
       simulator.run("onKeyDown", keyEvent("i"));
 
-      expect(simulator.value.focusText.text).to.equal("hello");
+      expect(simulator.value.focusText.text).toBe("hello");
     });
   });
 
@@ -69,7 +69,7 @@ describe("AutoReplacePlugin()", () => {
       simulator.run("onKeyDown", keyEvent("y"));
       simulator.run("onKeyDown", keyEvent("e"));
 
-      expect(simulator.value.focusText.text).to.equal("goodbye");
+      expect(simulator.value.focusText.text).toBe("goodbye");
     });
   });
 
@@ -83,7 +83,7 @@ describe("AutoReplacePlugin()", () => {
       simulator.run("onKeyDown", keyEvent("h"));
       simulator.run("onKeyDown", keyEvent("i"));
 
-      expect(simulator.value.focusText.text).to.equal("hello!");
+      expect(simulator.value.focusText.text).toBe("hello!");
     });
   });
 
@@ -107,28 +107,28 @@ describe("AutoReplacePlugin()", () => {
       // Write another character
       simulator.run("onKeyDown", keyEvent("!"));
 
-      expect(simulator.value.inlines.get(0).text).to.equal("hi!");
+      expect(simulator.value.inlines.get(0).text).toBe("hi!");
     });
   });
 
   describe("when writing a non printable character", () => {
     it("should ignore the event", () => {
       const value = initialValue;
-      const fakeTransform = fake();
+      const fakeTransform = jest.fn();
       const plugins = [AutoReplacePlugin(fakeTransform)];
       let simulator = new Editor({ value, plugins });
 
       // Press enter
       simulator.run("onKeyDown", keyEvent("Enter"));
 
-      expect(fakeTransform.callCount).to.equal(0);
+      expect(fakeTransform).toHaveBeenCalledTimes(0);
     });
   });
 
   describe("when writing ctrl or cmd character", () => {
     it("should ignore the event", () => {
       const value = initialValue;
-      const fakeTransform = fake();
+      const fakeTransform = jest.fn();
       const plugins = [AutoReplacePlugin(fakeTransform)];
       let simulator = new Editor({ value, plugins });
 
@@ -138,7 +138,7 @@ describe("AutoReplacePlugin()", () => {
       // Press cmd
       simulator.run("onKeyDown", cmdKeyEvent());
 
-      expect(fakeTransform.callCount).to.equal(0);
+      expect(fakeTransform).toHaveBeenCalledTimes(0);
     });
   });
 
@@ -155,7 +155,7 @@ describe("AutoReplacePlugin()", () => {
       // Write some text
       simulator.run("onKeyDown", keyEvent("h"));
 
-      expect(simulator.value.focusText.text).to.equal("h" + textToInsert);
+      expect(simulator.value.focusText.text).toBe("h" + textToInsert);
     });
   });
 });
