+++ package.json
@@ -14,9 +14,9 @@
     "slate": "^0.44.6"
   },
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
-    "babel-preset-env": "^1.7.0",
+    "@babel/core": "^7.7.0",
+    "@babel/preset-env": "^7.7.1",
+    "babel-jest": "^24.9.0",
     "immutable": "^3.8.2",
     "jest": "^24.9.0",
     "prettier": "^1.15.2",
